pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.12.0",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "streamlit>=1.50.0",
]
